Refactor command handling into separate files
