+++ handle_ping.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"fmt"
+	"io"
+	"time"
+)
+
+func handleVarnishCliPingRequest(writer io.Writer) {
+	response := fmt.Sprintf("PONG %d 1.0", time.Now().Unix())
+	writeVarnishCliResponse(writer, CLIS_OK, response)
+}
